Refactor test code - mocks.js
Put mock app and streambundle into its own file
